requirements/development.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=development.txt --strip-extras development.in
#
aiofiles==23.2.1
    # via -r production.txt
aiosqlite==0.19.0
    # via -r production.txt
appnope==0.1.3
    # via ipython
asttokens==2.4.1
    # via stack-data
autopep8==2.0.4
    # via -r lint.txt
black==23.11.0
    # via -r lint.txt
build==1.0.3
    # via
    #   -r lint.txt
    #   pip-tools
certifi==2023.11.17
    # via
    #   -r production.txt
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   -r lint.txt
    #   cryptography
cfgv==3.4.0
    # via
    #   -r lint.txt
    #   pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r lint.txt
    #   black
    #   pip-tools
codetiming==1.4.0
    # via -r development.in
coverage==7.3.2
    # via
    #   -r lint.txt
    #   pytest-cov
cryptography==41.0.5
    # via
    #   -r lint.txt
    #   trustme
    #   types-pyopenssl
    #   types-redis
decorator==5.1.1
    # via ipython
distlib==0.3.7
    # via
    #   -r lint.txt
    #   virtualenv
docopt==0.6.2
    # via -r production.txt
executing==2.0.1
    # via stack-data
factory-boy==3.3.0
    # via pytest-factoryboy
faker==20.0.0
    # via factory-boy
filelock==3.13.1
    # via
    #   -r lint.txt
    #   virtualenv
flake8==6.1.0
    # via -r lint.txt
identify==2.5.31
    # via
    #   -r lint.txt
    #   pre-commit
idna==3.4
    # via
    #   requests
    #   trustme
importlib-metadata==6.8.0
    # via
    #   -r lint.txt
    #   yapf
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipython==8.17.2
    # via -r development.in
isort==5.12.0
    # via -r lint.txt
jedi==0.19.1
    # via ipython
markdown-it-py==3.0.0
    # via rich
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via
    #   -r lint.txt
    #   flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.7.0
    # via -r lint.txt
mypy-extensions==1.0.0
    # via
    #   -r lint.txt
    #   black
    #   mypy
nodeenv==1.8.0
    # via
    #   -r lint.txt
    #   pre-commit
packaging==23.2
    # via
    #   -r lint.txt
    #   black
    #   build
    #   pytest
    #   pytest-sugar
parso==0.8.3
    # via jedi
pathspec==0.11.2
    # via
    #   -r lint.txt
    #   black
pexpect==4.8.0
    # via ipython
pip-tools==7.3.0
    # via -r lint.txt
platformdirs==3.11.0
    # via
    #   -r lint.txt
    #   black
    #   virtualenv
    #   yapf
pluggy==1.3.0
    # via pytest
pre-commit==3.5.0
    # via -r lint.txt
prompt-toolkit==3.0.40
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via
    #   -r lint.txt
    #   autopep8
    #   flake8
pycparser==2.21
    # via
    #   -r lint.txt
    #   cffi
pyflakes==3.1.0
    # via
    #   -r lint.txt
    #   flake8
pygments==2.16.1
    # via
    #   ipython
    #   rich
pyproject-hooks==1.0.0
    # via
    #   -r lint.txt
    #   build
pytest==7.4.3
    # via
    #   -r development.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-factoryboy
    #   pytest-mock
    #   pytest-sugar
pytest-asyncio==0.21.1
    # via -r development.in
pytest-cov==4.1.0
    # via pytest-cover
pytest-cover==3.0.0
    # via pytest-coverage
pytest-coverage==0.0
    # via -r development.in
pytest-factoryboy==2.6.0
    # via -r development.in
pytest-mock==3.12.0
    # via -r development.in
pytest-sugar==0.9.7
    # via -r development.in
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via -r production.txt
pytz==2023.3.post1
    # via -r production.txt
pyyaml==6.0.1
    # via
    #   -r lint.txt
    #   pre-commit
requests==2.31.0
    # via requests-mock
requests-mock==1.11.0
    # via -r development.in
rich==13.6.0
    # via -r development.in
ruff==0.1.5
    # via -r lint.txt
sentry-sdk==1.36.0
    # via -r production.txt
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
    #   requests-mock
stack-data==0.6.3
    # via ipython
termcolor==2.3.0
    # via pytest-sugar
tomli==2.0.1
    # via
    #   -r lint.txt
    #   yapf
traitlets==5.13.0
    # via
    #   ipython
    #   matplotlib-inline
trustme==1.1.0
    # via -r development.in
types-aiofiles==23.2.0.0
    # via -r lint.txt
types-cachetools==5.3.0.7
    # via -r lint.txt
types-pyopenssl==23.3.0.0
    # via
    #   -r lint.txt
    #   types-redis
types-pytz==2023.3.1.1
    # via -r lint.txt
types-redis==4.6.0.10
    # via -r lint.txt
types-requests==2.31.0.10
    # via -r lint.txt
types-setuptools==68.2.0.1
    # via -r lint.txt
typing-extensions==4.8.0
    # via
    #   -r lint.txt
    #   mypy
    #   pytest-factoryboy
urllib3==2.1.0
    # via
    #   -r lint.txt
    #   -r production.txt
    #   requests
    #   sentry-sdk
    #   types-requests
virtualenv==20.24.6
    # via
    #   -r lint.txt
    #   pre-commit
wcwidth==0.2.10
    # via prompt-toolkit
wheel==0.41.3
    # via
    #   -r lint.txt
    #   pip-tools
yapf==0.40.2
    # via -r lint.txt
zipp==3.17.0
    # via
    #   -r lint.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
